[IOTDB-1473] Fix the bug that run SessionExample.java will be failed in (#3770)
cluster mode. Root cause is auto create schema fail in cluster mode.
